requirements.txt
PyQt6
huggingface_hub
python-dotenv